refactor(edgeless): expose configurable tools
