+++ packages/blocks/src/index.ts
@@ -59,7 +59,7 @@ export type {
   TemplateCategory,
   TemplateManager,
 } from './root-block/edgeless/components/toolbar/template/template-type.js';
-export { CopilotSelectionController } from './root-block/edgeless/controllers/tools/copilot-tool.js';
+export { CopilotSelectionController } from './root-block/edgeless/tools/copilot-tool.js';
 export * from './root-block/index.js';
 export * from './schemas.js';
 export * from './specs/index.js';
