+++ packages/blocks/src/root-block/edgeless/components/panel/shape-panel.ts
@@ -4,7 +4,7 @@ import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
-import type { ShapeTool } from '../../controllers/tools/shape-tool.js';
+import type { ShapeTool } from '../../tools/shape-tool.js';
 
 import '../buttons/tool-icon-button.js';
 import { ShapeComponentConfig } from '../toolbar/shape/shape-menu-config.js';
