+++ packages/blocks/src/root-block/edgeless/components/toolbar/note/note-menu.ts
@@ -3,7 +3,7 @@ import { LitElement, css, html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
-import type { NoteTool } from '../../../controllers/tools/note-tool.js';
+import type { NoteTool } from '../../../tools/note-tool.js';
 import type { EdgelessTool } from '../../../types.js';
 
 import {
