+++ packages/blocks/src/root-block/edgeless/components/toolbar/shape/shape-draggable.ts
@@ -14,7 +14,7 @@ import { styleMap } from 'lit/directives/style-map.js';
 import type { DraggableShape } from './utils.js';
 
 import { CanvasElementType } from '../../../../../surface-block/index.js';
-import { ShapeToolController } from '../../../controllers/tools/shape-tool.js';
+import { ShapeToolController } from '../../../tools/shape-tool.js';
 import { EdgelessDraggableElementController } from '../common/draggable/draggable-element.controller.js';
 import { EdgelessToolbarToolMixin } from '../mixins/tool.mixin.js';
 import { buildVariablesObject } from './utils.js';
