+++ packages/blocks/src/root-block/edgeless/index.ts
@@ -1,6 +1,6 @@
 export { GfxBlockModel as EdgelessBlockModel } from './block-model.js';
 export { FramePreview } from './components/frame/frame-preview.js';
-export { EdgelessToolController } from './controllers/tools/index.js';
 export * from './edgeless-root-block.js';
 export { EdgelessRootService } from './edgeless-root-service.js';
+export { EdgelessToolController } from './tools/index.js';
 export * from './types.js';
