+++ packages/blocks/src/root-block/edgeless/services/tools-manager.ts
@@ -11,9 +11,9 @@ import { NoteDisplayMode } from '@blocksuite/affine-model';
 import { IS_MAC } from '@blocksuite/global/env';
 import { DisposableGroup } from '@blocksuite/global/utils';
 
-import type { EdgelessToolController } from '../controllers/tools/index.js';
 import type { EdgelessRootBlockComponent } from '../edgeless-root-block.js';
 import type { EdgelessRootService } from '../edgeless-root-service.js';
+import type { EdgelessToolController } from '../tools/index.js';
 import type { EdgelessTool } from '../types.js';
 import type { EdgelessSelectionState } from './selection-manager.js';
 
@@ -21,7 +21,7 @@ import {
   isMiddleButtonPressed,
   isRightButtonPressed,
 } from '../../../_common/utils/index.js';
-import { CopilotSelectionController } from '../controllers/tools/copilot-tool.js';
+import { CopilotSelectionController } from '../tools/copilot-tool.js';
 import { edgelessElementsBound } from '../utils/bound-utils.js';
 import { isNoteBlock } from '../utils/query.js';
