+++ packages/blocks/src/root-block/edgeless/tools/edgeless-tool.ts
@@ -1,10 +1,10 @@
 import type { PointerEventState } from '@blocksuite/block-std';
 
-import type { GfxBlockModel } from '../../block-model.js';
-import type { EdgelessRootBlockComponent } from '../../edgeless-root-block.js';
-import type { EdgelessRootService } from '../../edgeless-root-service.js';
-import type { SelectionArea } from '../../services/tools-manager.js';
-import type { EdgelessTool } from '../../types.js';
+import type { GfxBlockModel } from '../block-model.js';
+import type { EdgelessRootBlockComponent } from '../edgeless-root-block.js';
+import type { EdgelessRootService } from '../edgeless-root-service.js';
+import type { SelectionArea } from '../services/tools-manager.js';
+import type { EdgelessTool } from '../types.js';
 
 export abstract class EdgelessToolController<
   Tool extends EdgelessTool = EdgelessTool,
