+++ packages/blocks/src/root-block/edgeless/tools/pan-tool.ts
