+++ packages/blocks/src/root-block/edgeless/tools/template-tool.ts
