+++ packages/blocks/src/root-block/edgeless/utils/hotkey-utils.ts
@@ -1,7 +1,7 @@
 import { ShapeType } from '@blocksuite/affine-model';
 
-import type { ShapeTool } from '../controllers/tools/shape-tool.js';
 import type { EdgelessRootBlockComponent } from '../edgeless-root-block.js';
+import type { ShapeTool } from '../tools/shape-tool.js';
 
 const shapeMap: Record<ShapeTool['shapeType'], number> = {
   [ShapeType.Rect]: 0,
