+++ packages/blocks/src/root-block/widgets/edgeless-copilot-panel/toolbar-entry.ts
@@ -4,8 +4,8 @@ import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
 import type { AIItemGroupConfig } from '../../../_common/components/ai-item/types.js';
-import type { CopilotSelectionController } from '../../edgeless/controllers/tools/copilot-tool.js';
 import type { EdgelessRootBlockComponent } from '../../edgeless/edgeless-root-block.js';
+import type { CopilotSelectionController } from '../../edgeless/tools/copilot-tool.js';
 
 import { SurfaceGroupLikeModel } from '../../../surface-block/element-model/base.js';
 import { isFrameBlock } from '../../edgeless/utils/query.js';
